Added file matching using regex patterns
New regex file matching mode has been added to the renaming function and command line tool. This way, file names with
any kind of pattern including the recognized values (crc32, md5, sha1, title) can be renamed properly. For example, to
rename a file called "12245780 - foo.png" where the first 8 digits are the CRC32, we would use the rename tool with
crc as the source format and this regex parameter "-r (.*) - foo,1" where everything to the left of the comma is the
matching pattern itself and the number after it indicates which group we want to use. In this example just one group
(defined by parenthesis) is used but you can use as many groups as you want in your patterns.

For the renaming function, two new parameters have been defined, the pattern and the number of the group to capture.